Implemented reciprocal & encodeXCoord
reciprocal needs to be tested. encode XCoord appears to work as intended. There are some potential problems with the Number library you are using. Don't know if modulus works as intended. Might not be able to do a = a + b (using the output as an argument). Might not be able to do more than one operation at a time?