What is going on here!?
Found the most peculiar bug. When there are more than two block to hash, for whatever reason, initialize() fails to properly copy hInit to h. The first 64-bit word (and only the first 64-bit word), aka h[0], incorrectly copies in the final byte as 0x00 when it should be 0x08.