Chasing a phantom
I must be doing something wrong that is causing the program to be very fragile. hInit is not used anywhere in the program. Even so, the program gives a different answer depending on whether it is commented out or not. Currently, if it is commented out, a (the working variable) seems to get messed up at some point in hashProcess().