Did some research
I believe that if both a and b are < p, subtraction can be performed nearly exactly the same as addition, but with + p, rather than - p. Might even be faster as you may simply be able to use the negative sign to determine whether to add p or not, cutting out conditional searches.